fixed the delete provincia

+++ DataLayer/Migrations/EfCoreContextModelSnapshot.cs
@@ -29,7 +29,7 @@ namespace DataLayer.Migrations
 
                     b.Property<int?>("MaterialID");
 
-                    b.Property<decimal>("Precio");
+                    b.Property<double>("Precio");
 
                     b.HasKey("AccionC_MaterialID");
 
@@ -90,7 +90,7 @@ namespace DataLayer.Migrations
 
                     b.Property<int>("Cantidad");
 
-                    b.Property<decimal>("Precio");
+                    b.Property<double>("Precio");
 
                     b.Property<string>("UnidadMedida");
 
@@ -139,7 +139,7 @@ namespace DataLayer.Migrations
 
                     b.Property<int>("Año");
 
-                    b.Property<decimal>("Presupuesto");
+                    b.Property<double>("Presupuesto");
 
                     b.Property<string>("TipoPlan");
